Move comparison demo to dedicated route
