+++ docs/app/compare/_components/comparison-mode-controller.tsx
