+++ docs/app/compare/_components/use-has-mounted.ts
@@ -1,0 +1,12 @@
+"use client";
+
+import { useSyncExternalStore } from "react";
+
+const subscribe = () => () => {};
+const getClientSnapshot = () => true;
+const getServerSnapshot = () => false;
+
+/** Returns false during SSR and the hydration pass, then true on the client. */
+export function useHasMounted() {
+  return useSyncExternalStore(subscribe, getClientSnapshot, getServerSnapshot);
+}
