+++ packages/react-ui/src/components/OpenUIChat/index.ts
@@ -1,5 +1,3 @@
-export { GenUIAssistantMessage } from "./GenUIAssistantMessage";
-export type { GenUIAssistantMessageProps, GenUIConversationAction } from "./GenUIAssistantMessage";
 export { GenUIUserMessage } from "./GenUIUserMessage";
 export { ShareThread } from "./ShareThread";
 export type { ShareThreadProps } from "./ShareThread";
